Delays interrupt enabling for initialization
Changes the interrupt initialization process to delay enabling
the interrupt until after the device initialization is complete.
This ensures that the interrupt is not triggered prematurely
before the device is fully configured. It also moves the interrupt
initialization to `main.c` so all drivers can be initialized first.

Also reorders source entries in .cproject to be more consistent.